(maint) Update integer identifiers to be long
This updates all of the identifiers for comments and issues to use
a long instead of an integer.

This is done as reading or submitting comments was throwing an
exception due to the number being too large for integers.
As such these needed to change, the other identifiers is also changed
for future proofing so we don't need to change these in the future.